removed old code and fixed some warnings

+++ core/base/mergeTreeClustering/BranchMappingDistance.h
@@ -257,22 +257,6 @@ namespace ttk {
       std::vector<std::tuple<ftm::idNode, ftm::idNode, double>> *outputMatching
       = nullptr) {
 
-      // // optional preprocessing
-      // if(preprocess_ && !writeOptimalBranchDecomposition_){
-      //   treesNodeCorr_.resize(2);
-      //   preprocessingPipeline<dataType>(
-      //     mTree1, epsilonTree1_, epsilon2Tree1_, epsilon3Tree1_,
-      //     branchDecomposition_, useMinMaxPair_, cleanTree_,
-      //     treesNodeCorr_[0],true,true);
-      //   preprocessingPipeline<dataType>(
-      //     mTree2, epsilonTree2_, epsilon2Tree2_, epsilon3Tree2_,
-      //     branchDecomposition_, useMinMaxPair_, cleanTree_,
-      //     treesNodeCorr_[1],true,true);
-      // }
-
-      // ftm::FTMTree_MT *tree1 = (&mTree1.tree);
-      // ftm::FTMTree_MT *tree2 = (&mTree2.tree);
-
       // compute preorder of both trees (necessary for bottom-up dynamic
       // programming)
 
@@ -660,32 +644,6 @@ namespace ttk {
             linkedNodes2[m.second.first] = m.second.second;
             linkedNodes2[m.second.second] = m.second.first;
           }
-          // dataType cost = this->baseMetric_ == 0 ?
-          // editCost_Wasserstein1<dataType>(
-          //                   m.first.first, m.first.second, m.second.first,
-          //                   m.second.second, tree1, tree2)
-          //                 : this->baseMetric_ == 1 ?
-          //                 editCost_Wasserstein2<dataType>(
-          //                     m.first.first, m.first.second, m.second.first,
-          //                     m.second.second, tree1, tree2)
-          //                 : this->baseMetric_ == 2
-          //                   ? editCost_Persistence<dataType>(
-          //                     m.first.first, m.first.second, m.second.first,
-          //                     m.second.second, tree1, tree2)
-          //                   : editCost_Shifting<dataType>(
-          //                     m.first.first, m.first.second, m.second.first,
-          //                     m.second.second, tree1, tree2);
-          // dataType cost_ = editCost_Wasserstein1<dataType>(
-          //                   m.first.first, m.first.second, m.second.first,
-          //                   m.second.second, tree1, tree2);
-          // cost_mapping += cost_;
-          // std::cout << "(" << m.first.first << " " << m.first.second << ") -
-          // (" << m.second.first << " " << m.second.second << ") : " << cost <<
-          // " " << cost_;// << std::endl; std::cout << ";        (" <<
-          // tree1->getValue<dataType>(m.first.first) << " " <<
-          // tree1->getValue<dataType>(m.first.second) << ") - (" <<
-          // tree2->getValue<dataType>(m.second.first) << " " <<
-          // tree2->getValue<dataType>(m.second.second) << ")" << std::endl;
           if(m.first.first == -1)
             continue;
           if(m.first.second == -1)
@@ -710,25 +668,15 @@ namespace ttk {
                 : editCost_Shifting<dataType>(m.first.first, m.first.second,
                                               m.second.first, m.second.second,
                                               tree1, tree2);
-          if(m.first.second == tree1->getRoot()) {
+          if(m.first.second == (int)tree1->getRoot()) {
             matchedCost[m.first.second] = matchedCost[m.first.first];
           }
         }
-        // std::cout << "Pairs Tree 1:\n";
-        // for(int i=0; i<linkedNodes1.size(); i++){
-        //   std::cout << i << ": " << linkedNodes1[i] << std::endl;
-        // }
-        // std::cout << "Pairs Tree 2:\n";
-        // for(int i=0; i<linkedNodes2.size(); i++){
-        //   std::cout << i << ": " << linkedNodes2[i] << std::endl;
-        // }
         for(ftm::idNode i = 0; i < matchedNodes.size(); i++) {
           if(matchedNodes[i] >= 0)
             outputMatching->emplace_back(
               std::make_tuple(i, matchedNodes[i], matchedCost[i]));
         }
-
-        // std::cout << res << " " << cost_mapping << std::endl;
       }
 
       return squared_ ? std::sqrt(res) : res;
